requirements.txt
sqlalchemy
fastapi
pydantic
pytest
httpx
jupyter
matplotlib
uvicorn
python-dotenv